### 2024-09-16 #7235 @editorialbot
[PRE REVIEW]: Mashpit: sketching out genomic epidemiology
**Submitting author:** @tongzhouxu (<a href="http://orcid.org/0000-0002-3829-8688">Tongzhou Xu</a>)
**Repository:** https://github.com/tongzhouxu/mashpit
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.7
