+++ paper/paper.bib
@@ -1,4 +1,5 @@
 @article{hadfield2018nextstrain,
+  doi = {10.1093/bioinformatics/bty407},
   title={Nextstrain: real-time tracking of pathogen evolution},
   author={Hadfield, James and Megill, Colin and Bell, Sidney M and Huddleston, John and Potter, Barney and Callender, Charlton and Sagulenko, Pavel and Bedford, Trevor and Neher, Richard A},
   journal={Bioinformatics},
@@ -10,6 +11,7 @@
 }
 
 @article{nadon2017pulsenet,
+  doi = {10.2807/1560-7917.es.2017.22.23.30544},
   title={PulseNet International: Vision for the implementation of whole genome sequencing (WGS) for global food-borne disease surveillance},
   author={Nadon, Celine and Van Walle, Ivo and Gerner-Smidt, Peter and Campos, Josefina and Chinen, Isabel and Concepcion-Acevedo, Jeniffer and Gilpin, Brent and Smith, Anthony M and Kam, Kai Man and Perez, Enrique and others},
   journal={Eurosurveillance},
@@ -21,6 +23,7 @@
 }
 
 @article{ondov2016mash,
+  doi = {10.1186/s13059-016-0997-x},
   title={Mash: fast genome and metagenome distance estimation using MinHash},
   author={Ondov, Brian D and Treangen, Todd J and Melsted, P{\'a}ll and Mallonee, Adam B and Bergman, Nicholas H and Koren, Sergey and Phillippy, Adam M},
   journal={Genome biology},
@@ -45,6 +48,7 @@
 }
 
 @article{camacho2009blast,
+  doi = {10.1186/1471-2105-10-421},
   title={BLAST+: architecture and applications},
   author={Camacho, Christiam and Coulouris, George and Avagyan, Vahram and Ma, Ning and Papadopoulos, Jason and Bealer, Kevin and Madden, Thomas L},
   journal={BMC bioinformatics},
@@ -56,6 +60,7 @@
 }
 
 @article{zhang2019seqsero2,
+  doi = {10.1128/aem.01746-19},
   title={SeqSero2: rapid and improved Salmonella serotype determination using whole-genome sequencing data},
   author={Zhang, Shaokang and den Bakker, Hendrik C and Li, Shaoting and Chen, Jessica and Dinsmore, Blake A and Lane, Charlotte and Lauer, AC and Fields, Patricia I and Deng, Xiangyu},
   journal={Applied and environmental microbiology},
@@ -67,6 +72,7 @@
 }
 
 @article{zhou2020enterobase,
+  doi = {10.1101/gr.251678.119},
   title={The EnteroBase user's guide, with case studies on Salmonella transmissions, Yersinia pestis phylogeny, and Escherichia core genomic diversity},
   author={Zhou, Zhemin and Alikhan, Nabil-Fareed and Mohamed, Khaled and Fan, Yulei and Achtman, Mark and Brown, Derek and Chattaway, Marie and Dallman, Tim and Delahay, Richard and Kornschober, Christian and others},
   journal={Genome research},
@@ -78,6 +84,7 @@
 }
 
 @article{katz2017comparative,
+  doi = {10.3389/fmicb.2017.00375},
   title={A comparative analysis of the Lyve-SET phylogenomics pipeline for genomic epidemiology of foodborne pathogens},
   author={Katz, Lee S and Griswold, Taylor and Williams-Newkirk, Amanda J and Wagner, Darlene and Petkau, Aaron and Sieffert, Cameron and Van Domselaar, Gary and Deng, Xiangyu and Carleton, Heather A},
   journal={Frontiers in Microbiology},
