+++ mashpit/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'Tongzhou Xu'
 __email__ = 'tongzhou.xu@uga.edu'
-__version__ = '0.9.6'
+__version__ = '0.9.7'
