### 2024-09-20 #7235 @Kevin-Mattheus-Moerman
- [ ] In the paper, please spell out `USA` as `United States of America`.
