### 2024-09-20 #7235 @tongzhouxu
> @tongzhouxu Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:
>
> * [ ]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).
> * [ ]  In the paper, please spell out `USA` as `United States of America`.

Thank you for the suggestions. I have added a CONTRIBUTING.md in the latest commit.
