### 2024-09-28 #7235 @csoneson
👋🏻 @hkaspersen, @mberacochea, @ns-rse - would you be interested in reviewing this submission for JOSS?

> Mashpit: sketching out genomic epidemiology
https://github.com/openjournals/joss-reviews/issues/7235

Thanks in advance!
