### 2024-09-30 #7235 @csoneson
Hi @hkaspersen and @mberacochea - thanks a lot for accepting to review! I will assign you and start the review issue.

@mberacochea - next week is totally fine!
@hkaspersen - you can find more information [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). I'm starting the review issue now, and in there you will find more practical instructions, including how to create your checklist.
