### 2024-10-11 #7306 @hkaspersen
I have now completed my review of this software and manuscript. The authors present an interesting and useful tool that will solve a lot of issues regarding sensitive data and analysis speed. Below are some specific comments I would like the authors to address.

**Functionality**:
The functionality of the software (v. 0.9.7) was tested on our HPC cluster, using Miniconda for installation by following the guidelines from the GitHub page. No problems were encountered during installation. The software was tested by sketching the *Salmonella* database as described in the example commands, followed by querying a local *Salmonella* genome against the database, using default settings.
- [ ] The following warning was encountered: `FutureWarning: Minimal version of pyarrow will soon be increased to 14.0.1. You are using 9.0.0. Please consider upgrading.`. Unless version 9.0.0 of `pyarrow` is crucial, please consider updating this software.
- [ ] The time used to prepare the *Salmonella* database was quite long, which is understandable. However, it would be useful for users to know how long one can expect this preparation to take, so they can plan accordingly.
- [ ] Output: The `png` tree figure is clipped due to the number of genomes included (default 200). Make sure the figure is within the limits of the plot, or state in the documentation that "no tree figure will be generated when including more than x results".
- [ ] I cannot reproduce the performance evaluation of Mashpit from the description provided in the manuscript, due to missing results description (see comment under "Manuscript")

**Documentation**:
The documentation provided on GitHub is a bit lacking and could be expanded.
- [ ] Please provide a longer description of the software and its uses on the github documentation. This makes it easier for users to understand what the software does and why it is useful.
- [ ] No automated tests or test data is provided. Please consider adding a small dataset for users to test the software with. No github actions with CI tests was identified in the github repository.

**Manuscript**:
The manuscript is well-written and provides context on why the software is useful, and what problem it solves.
- Summary:
    - Line 28 - 34: A brief description is provided on the performance of Mashpit compared to BLAST, but no actual comparison is presented. Please consider adding data on how much faster and/or accurate Mashpit is compared to BLAST. Are there any other software worth comparing to?
- Mashpit design:
    - [ ] Line 65 - 68: The authors state that they have evaluated the performance of Mashpit on four different foodborne pathogens. However, the results from these tests are not presented clearly. I assume this is the data presented in Figure 1? If so, please refer to the figure here. Due to the figure being clipped in the pdf, it makes it difficult to grasp all the details in the figure. Additionally, it would be very beneficial for the reader to know the exact results from these tests, and adding them to the text would be useful.
- Minor comments:
    - [ ] Line 43: Please Italicise *Salmonella*
    - [ ] Line 68: Please correct *Campylobactor* to *Campylobacter*
    - [ ] Line 78: Add 'h' to 'Mashpit'
    - [ ] Line 80: Correct 'calclated'
    - [ ] Figure 1: Only parts of the figure is visible. Make sure the full figure is visible.
