### 2024-11-09 tongzhouxu/mashpit#26 @mberacochea
[JOSS REVIEW #7306] Warning pyarrow in mashpit built
Hi @tongzhouxu

Not a massive problem, but it could be confusing for users. Version 0.9.6 prints "FutureWarning: Minimal version of pyarrow will soon be increased to 14.0.1. You are using 9.0.0. Please consider upgrading." when using the `build` and `query` commands.

https://github.com/openjournals/joss-reviews/issues/7306

Cheers
