### 2024-11-10 tongzhouxu/mashpit#27 @mberacochea
[JOSS REVIEW #7306] Webserver not working
Hi,

I'm trying to run the webserver but I'm getting the following error:
```
$ mashpit webserver
/home/mbc/software/micromamba/envs/envs/mashpit/lib/python3.9/site-packages/dask/dataframe/_pyarrow_compat.py:15: FutureWarning: Minimal version of pyarrow will soon be increased to 14.0.1. You are using 9.0.0. Please consider upgrading.
  warnings.warn(
Traceback (most recent call last):
  File "/home/mbc/software/micromamba/envs/envs/mashpit/bin/mashpit", line 10, in <module>
    sys.exit(main())
  File "/home/mbc/software/micromamba/envs/envs/mashpit/lib/python3.9/site-packages/mashpit/mashpit.py", line 69, in main
    args.func(args)
AttributeError: 'Namespace' object has no attribute 'func'
```

I'm running this using python 3.9, on Ubuntu (using WSL) - not sure if this could be a problem.

https://github.com/openjournals/joss-reviews/issues/7306

Cheers
