### 2024-11-10 tongzhouxu/mashpit#28 @mberacochea
[JOSS REVIEW #7306] Expand the documentation
Hi,

I was able to run the install and run the tool by following the instructions. But, I think the tool would benefit from a slighly expanded README. These following is a list of the improvements I suggest:

- [x] Include a slighly longer description of the tool and it's purpose
- [x] Use a "smaller" example, Salmonella is massive and it takes a really long time to run the `build` command. This would likely to put off users if they want to get a sense of what the tool is capable of. Legionella_pneumophila (which I used to test it and worked perfectly on my laptop)
- [x] Inculde a link to a genome or sample to test the `query` command
- [x] Include a section in README explaning what are the outputs and describing the contents of them
- [x] Expand on the webserver, what is the purpose of the webserver? Maybe include a few screencaps too.
- [x] Tests, I was able to run the tests by reading the github actions workflow definition, please consider including this also the README (how to run the unit tests)

https://github.com/openjournals/joss-reviews/issues/7306

Cheers
