### 2024-11-10 tongzhouxu/mashpit#29 @mberacochea
[JOSS REVIEW #7306] Pin the dependencies 
Hi,

The setup.py file includes the tool dependencies, but they are not pinned to specific versions. Please pin the versions, as this will improve the stability of the packages; new versions of the dependencies may break Mashpit.

Additionally, Python has been gradually moving from setup.py to pyproject.toml as part of the packaging revamp. I would suggest migrating to pyproject.toml (https://packaging.python.org/en/latest/tutorials/packaging-projects/).

https://github.com/openjournals/joss-reviews/issues/7306

Cheers
