### 2024-11-10 tongzhouxu/mashpit#30 @mberacochea
[JOSS REVIEW #7306] Code style formatting
Hi,

This is a suggestion, not mandatory for JOSS. The code is well written and easy to follow—kudos for that. Using [Black](https://github.com/psf/black) or [Ruff](https://github.com/astral-sh/ruff) to format the code would enhance its readability at basically no cost. Please consider formatting the code with Black.

https://github.com/openjournals/joss-reviews/issues/7306.

Cheers
