### 2024-11-10 tongzhouxu/mashpit#31 @mberacochea
 [JOSS REVIEW #7306] Unit tests for the query functionality
Hi,

The test suite is pretty good; it covers much of the tool's internal API. However, if I'm not mistaken, it doesn’t cover mashpit.query, which is a key component of the tool. Could you consider adding a few unit tests for it? This would help make the app more robust.

https://github.com/openjournals/joss-reviews/issues/7306

Cheers
