### 2024-11-10 #7306 @mberacochea
I've finished my review :).

The authors present a command-line utility that addresses a problem in a convenient and performant way. The code is well-structured, and the installation and usage instructions work as expected. The repository includes a set of unit tests covering a significant portion of the API, automated through GitHub Actions. This approach leverages the efficient use of Sourmash to compare large numbers of genomes very quickly, providing a highly convenient tool that reduces the friction of this approach—from downloading references to querying using the user’s sample FASTA file.

## Source code - docs and functionality

I've submitted a series of issues with my suggestions to improve the source code, those are:

- [ ] https://github.com/tongzhouxu/mashpit/issues/26
- [ ] https://github.com/tongzhouxu/mashpit/issues/27
- [ ] https://github.com/tongzhouxu/mashpit/issues/28
- [ ] https://github.com/tongzhouxu/mashpit/issues/29
- [ ] https://github.com/tongzhouxu/mashpit/issues/30
- [ ] https://github.com/tongzhouxu/mashpit/issues/31

These issues cover aspects of the quality of the source, functionality and documentation. I'll update my check list after the author review my tickets in the repo.

## Paper

The paper is well written and it allows the reader to understand the purpose and scope of the application.

Some notes on particular lines:
- Lines 11 and 12. PulseNet should be cited on line 11, and maybe mention that is hosted by NCBI.
- Line 43, _Salmonella_ should be italicise
- Line 53, is sourmash an interface for Mash, or an implementation of the FracMinHash algorithm instead?
- Line 80, typo "calclated"
- Figure 1, it's clipped and not readable.
- The webserver it's not mentioned in the paper.

A more general note about Mashpit paper:
If I understood correctly, the accuracy of placing a genome in the correct SNP cluster isn’t very high (70% for _Salmonella_ when considering that the "right" SNP cluster should be within the top 25). This seems quite relevant, as it’s likely important to users (this is my assumption). I would suggest expanding or rephrasing this part to make it clearer to users what the best use case for Mashpit is (it’s already mentioned in the discussion, but I feel it’s a bit lost there). I would also suggest splitting Figure 1 into two figures, one to explain the "compute performance" (in terms of resources) and the other to show the tool's accuracy.
