### 2024-11-29 4256230
Address JOSS reviewers suggestions for documentation and testing
