+++ paper/paper.bib
@@ -94,4 +94,12 @@
   publisher={Frontiers}
 }
 
-
+@article{topaz2018bmscan,
+  title={BMScan: using whole genome similarity to rapidly and accurately identify bacterial meningitis causing species},
+  author={Topaz, Nadav and Boxrud, Dave and Retchless, Adam C and Nichols, Megan and Chang, How-Yi and Hu, Fang and Wang, Xin},
+  journal={BMC Infectious Diseases},
+  volume={18},
+  pages={1--9},
+  year={2018},
+  publisher={Springer}
+}
