+++ src/mashpit/query.py
@@ -1,0 +1,246 @@
+#!/usr/bin/env python3
+import os
+import glob
+import logging
+import ntpath
+import screed
+import sourmash
+import heapq
+import time
+import pandas as pd
+
+from skbio import DistanceMatrix
+from skbio.tree import nj
+from phytreeviz import TreeViz
+from mashpit.build import create_connection
+from operator import itemgetter
+from sourmash import (
+    SourmashSignature,
+    save_signatures,
+    load_one_signature,
+    load_file_as_signatures,
+)
+
+
+def get_query_sig(query_path, query_name, hash_number, kmer_size):
+    genome = query_path
+    mh = sourmash.MinHash(n=hash_number, ksize=kmer_size)
+    for record in screed.open(genome):
+        mh.add_sequence(record.sequence, True)
+    sig = SourmashSignature(mh, name=genome)
+    query_sig_path = os.path.join(os.path.dirname(query_path), f"{query_name}.sig")
+    with open(query_sig_path, "wt") as fp:
+        save_signatures([sig], fp)
+
+
+def generate_query_table(conn, sorted_asm_similarity_dict):
+    c = conn.cursor()
+    c.execute("SELECT * FROM METADATA")
+    names = [description[0] for description in c.description]
+
+    # List to collect DataFrames
+    df_list = []
+
+    for i in sorted_asm_similarity_dict:
+        sql_query = pd.read_sql_query(
+            f"SELECT * FROM METADATA WHERE asm_acc = '{i}'", conn
+        )
+        df_query = pd.DataFrame(sql_query, columns=names)
+        df_query["similarity_score"] = sorted_asm_similarity_dict[i]
+        df_list.append(df_query)
+
+    # Concatenate all DataFrames at once
+    output_df = pd.concat(df_list, ignore_index=True)
+
+    # Additional processing (if it is a taxon database, add the link of the SNP cluster to the output)
+    c.execute("SELECT value FROM DESC WHERE name = 'Type';")
+    db_type = c.fetchone()[0]
+    if db_type == "Taxonomy":
+        pds_list = output_df["PDS_acc"].to_list()
+        cluster_link = [
+            f"https://www.ncbi.nlm.nih.gov/pathogens/isolates/#{pds}"
+            for pds in pds_list
+        ]
+        output_df["SNP_tree_link"] = cluster_link
+
+    return output_df
+
+
+def generate_mashtree(
+    output_df, min_similarity, query_name, sig_path, added_annotation, database_sig
+):
+    # check if the top query similarity is smaller than the threshold
+    if float(output_df["similarity_score"].iloc[0]) < min_similarity:
+        logging.error(
+            "Top query similarity is smaller than the threshold. Mashtree can not be generated. "
+        )
+        exit(1)
+    # select top results that are above the threshold
+    acc_list = output_df[output_df["similarity_score"] >= min_similarity][
+        "asm_acc"
+    ].to_list()
+    # if the number of top results is smaller than 2, mashtree can not be generated
+    if len(acc_list) < 2:
+        logging.error(
+            "Number of top results is smaller than 2. Mashtree can not be generated. "
+        )
+        exit(1)
+    leaves = [query_name] + acc_list
+    # select database signatures that are present in acc_list
+    sigs = []
+    for sig in database_sig:
+        if sig.name in acc_list:
+            sigs.append(sig)
+    # sort sigs based on the order of acc_list
+    sigs = sorted(sigs, key=lambda x: acc_list.index(x.name))
+    matrix = []
+    # first row
+    first_row = [0]
+    for similarity_score in output_df[output_df["similarity_score"] >= 0.85][
+        "similarity_score"
+    ].to_list():
[…]
