+++ src/mashpit/__init__.py
@@ -1,0 +1,3 @@
+__author__ = "Tongzhou Xu"
+__email__ = "tongzhou.xu@uga.edu"
+__version__ = "0.9.8"

[…]
