+++ .github/workflows/python-app.yml
@@ -27,6 +27,8 @@ jobs:
           echo "$HOME/.local/bin" >> $GITHUB_PATH
       - name: Display Python version
         run: python -c "import sys; print(sys.version)"
+      - name: Install HDF5
+        run: sudo apt-get update && sudo apt-get install -y libhdf5-dev
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
