+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools"]
+requires = ["setuptools ~= 75.1.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -16,6 +16,7 @@ requires-python = ">=3.8"
 dependencies = [
     "sourmash ~= 4.6.1",
     "pandas ~= 2.0.3",
+    "numpy ~= 1.24.4",
     "biopython ~= 1.83",
     "scikit-bio ~= 0.6.2",
     "tqdm ~= 4.67.0",
