+++ src/mashpit/build.py
@@ -630,6 +630,8 @@ def build_accession(args):
         if len(esearch_record["IdList"]) == 0:
             continue
         # get assembly accession
+        if args.key == None:
+            time.sleep(1)
         esummary_handle = Entrez.esummary(
             db="assembly", id=esearch_record["IdList"][0], report="full"
         )
@@ -639,8 +641,13 @@ def build_accession(args):
         ]
         gca_acc_list.append(asm_acc)
         # get metadata
+                # if api key is not provided, wait for 1 second
+        if args.key == None:
+            time.sleep(1)
         esearch_handle = Entrez.esearch(db="biosample", term=biosample, retmax="3")
         esearch_record = Entrez.read(esearch_handle)
+        if args.key == None:
+            time.sleep(1)
         efetch_handle = Entrez.efetch(db="biosample", id=esearch_record["IdList"][0])
         efetch_record = efetch_handle.read()
         root = ET.fromstring(efetch_record)
