+++ paper/paper.bib
@@ -103,3 +103,25 @@
   year={2018},
   publisher={Springer}
 }
+
+@article{swaminathan2001pulsenet,
+  title={PulseNet: the molecular subtyping network for foodborne bacterial disease surveillance, United States.},
+  author={Swaminathan, Bala and Barrett, Timothy J and Hunter, Susan B and Tauxe, Robert V and Force, CDC PulseNet Task},
+  journal={Emerging infectious diseases},
+  volume={7},
+  number={3},
+  pages={382},
+  year={2001},
+  publisher={Centers for Disease Control and Prevention}
+}
+
+@article{shu2017gisaid,
+  title={GISAID: Global initiative on sharing all influenza data--from vision to reality},
+  author={Shu, Yuelong and McCauley, John},
+  journal={Eurosurveillance},
+  volume={22},
+  number={13},
+  pages={30494},
+  year={2017},
+  publisher={European Centre for Disease Prevention and Control}
+}
