+++ paper/paper.md
@@ -2,7 +2,7 @@
 title: 'Mashpit: sketching out genomic epidemiology'
 tags:
   - Python
-  - Min-hash 
+  - MinHash 
   - Mash
   - Sourmash
   - Outbreak
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 We are in the era of genomic epidemiology.
 The surveillance of many transmissible diseases is increasingly being conducted through whole genome sequencing of pathogenic agents. 
-One notable example is _Salmonella_, a major foodborne pathogen routinely sequenced by surveillance programs such as PulseNet. 
+One notable example is _Salmonella_, a major foodborne pathogen routinely sequenced by surveillance programs such as PulseNet [@swaminathan2001pulsenet]. 
 Large volumes of _Salmonella_ genomes from these programs are deposited in database systems including NCBI [@nadon2017pulsenet]. 
 These publicly available genomes can be analyzed in a variety of ways such as serotyping [@zhang2019seqsero2],
 multilocus sequence typing (MLST) [@zhou2020enterobase], and single nucleotide polymorphism (SNP) typing [@katz2017comparative].
@@ -49,8 +49,8 @@ About once a day, it compares all genomes of a given taxon, separates all genome
 This method is quite comprehensive, but it relies on each sample being public, and it cannot be executed locally.
 
 Another approach is to provide new tools for decentralized and customized manipulation of genomics resources.
-We observed that an algorithm for genomics called Min-Hash is well positioned for this purpose.
-A commonly used software for Min-Hash is called Mash [@ondov2016mash].
+We observed that an algorithm for genomics called MinHash is well positioned for this purpose.
+A commonly used software for MinHash is called Mash [@ondov2016mash].
 Querying with Mash can be about 4 orders of magnitude faster than other common methods like Basic Local Alignment Search Tool (BLAST) and can have a smaller disk footprint [@camacho2009blast;@topaz2018bmscan].
 Therefore it can be run on more common scientific workstations.
 
@@ -59,8 +59,8 @@ We present Mashpit, a new rapid genomic epidemiology platform to query against t
 # Statement of need 
 
 Querying a sample against these magnitudes of genomes is becoming less sustainable, especially for smaller laboratories.
-Currently, GISAID and NCBI are staying ahead of the curve by producing a global tree of each organism every day.
-This requires herculean efforts, cutting edge algorithms, and powerful computers.
+Currently, GISAID and NCBI are staying ahead of the curve by producing a global tree of each organism every day [@shu2017gisaid].
+This requires herculean efforts, cutting-edge algorithms, and powerful computers.
 However, smaller laboratories usually have a scientific workstation or similar equipment, much different than a cluster computing system.
 
 We note that for some organisms like _Salmonella_, queries can be of a sensitive nature.
@@ -71,7 +71,7 @@ Mashpit queries genomes locally using Mash, thereby achieving speedy results whi
 
 # Mashpit design
 
-Mashpit is comprised of three major parts: A min-hash database, its associated metadata, and the min-hash querying.
+Mashpit is comprised of three major parts: A MinHash database, its associated metadata, and the MinHash querying.
 
 The database is created with an interface to Mash, called Sourmash [@Brown2016].
 Each genome is imported by sketching it and adding it to a Sourmash signature database.
@@ -83,26 +83,38 @@ where n is the number of genomes in the cluster.
 The centroid genome $g_c$ is calculated as:
 $$g_c=\underset{g_i∈G}{argmin}\sum_{j=1}^{n} d(g_i,g_j)$$
 Where $d(g_i,g_j)$ is the distance between two genomes.
-By default, Mashpit will download the latest SNP cluster for specified species and uses a kmer size of 31 and kmer number of 1000 for sketching the genomes. 
-
[…]
