### 2024-12-03 tongzhouxu/mashpit#30 @tongzhouxu
Mashpit now uses Black for code formatting
