+++ paper/paper.bib
@@ -125,3 +125,13 @@
   year={2017},
   publisher={European Centre for Disease Prevention and Control}
 }
+@article{jain2018high,
+  title={High throughput ANI analysis of 90K prokaryotic genomes reveals clear species boundaries},
+  author={Jain, Chirag and Rodriguez-R, Luis M and Phillippy, Adam M and Konstantinidis, Konstantinos T and Aluru, Srinivas},
+  journal={Nature communications},
+  volume={9},
+  number={1},
+  pages={5114},
+  year={2018},
+  publisher={Nature Publishing Group UK London}
+}
