+++ .github/workflows/python-publish.yml
@@ -36,7 +36,12 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: release-dists
-          path: dist/
+          path: mashpit/dist/
+      - name: troubleshooting
+        run: |
+          pwd
+          ls -la
+          ls -la dist
 
   pypi-publish:
     runs-on: ubuntu-latest
