+++ pyproject.toml
@@ -12,6 +12,7 @@ authors = [
 license = {file = "LICENSE"}
 keywords = ["sourmash","surveillance","pathogen"]
 requires-python = ">=3.8"
+readme = "README.md"
 
 dependencies = [
     "sourmash ~= 4.6.1",
