### 2024-12-07 tongzhouxu/mashpit#29 @tongzhouxu
Mashpit is converted to pyproject.toml in v0.9.8. The dependencies were pinned in the pyproject.toml
