### 2024-12-07 tongzhouxu/mashpit#28 @tongzhouxu
The documentation is now expanded in v0.9.8
