### 2024-12-07 tongzhouxu/mashpit#27 @tongzhouxu
A unit test for webserver was added in v0.9.8. Please let me know if the problem persists.
