### 2024-12-07 tongzhouxu/mashpit#26 @tongzhouxu
This warning should be solved in v0.9.8
