### 2024-12-07 #7306 @tongzhouxu
Hi @mberacochea, thank you for all the suggestions. We have addressed all the points raised. Please find our detailed responses below:

> * [x]  The following warning was encountered: `FutureWarning: Minimal version of pyarrow will soon be increased to 14.0.1. You are using 9.0.0. Please consider upgrading.`. Unless version 9.0.0 of `pyarrow` is crucial, please consider updating this software.

This warning has been resolved in v0.9.8.

> * [x]  The time used to prepare the _Salmonella_ database was quite long, which is understandable. However, it would be useful for users to know how long one can expect this preparation to take, so they can plan accordingly.

We have added details on the database preparation time to the documentation.

> * [x]  Output: The `png` tree figure is clipped due to the number of genomes included (default 200). Make sure the figure is within the limits of the plot, or state in the documentation that "no tree figure will be generated when including more than x results".

This issue has been resolved in v0.9.8.

> * [x]  I cannot reproduce the performance evaluation of Mashpit from the description provided in the manuscript, due to missing results description (see comment under "Manuscript")

We have modified the figures and expanded the description of the performance evaluation.

> **Documentation**: The documentation provided on GitHub is a bit lacking and could be expanded.
> * [x]  Please provide a longer description of the software and its uses on the github documentation. This makes it easier for users to understand what the software does and why it is useful.

We have expanded the documentation in v0.9.8.

> * [x]  No automated tests or test data is provided. Please consider adding a small dataset for users to test the software with. No github actions with CI tests was identified in the github repository.

We use github action for the unit testing.

> **Manuscript**: The manuscript is well-written and provides context on why the software is useful, and what problem it solves.
>
> * Summary:
>
>   * Line 28 - 34: A brief description is provided on the performance of Mashpit compared to BLAST, but no actual comparison is presented. Please consider adding data on how much faster and/or accurate Mashpit is compared to BLAST. Are there any other software worth comparing to?

We have added a reference for the comparison between BLAST and Mash.

> * Mashpit design:
>
>   * [x]  Line 65 - 68: The authors state that they have evaluated the performance of Mashpit on four different foodborne pathogens. However, the results from these tests are not presented clearly. I assume this is the data presented in Figure 1? If so, please refer to the figure here. Due to the figure being clipped in the pdf, it makes it difficult to grasp all the details in the figure. Additionally, it would be very beneficial for the reader to know the exact results from these tests, and adding them to the text would be useful.

We have modified the figures and expanded the description of the performance evaluation.

> * Minor comments:
>
>   * [x]  Line 43: Please Italicise _Salmonella_
>   * [x]  Line 68: Please correct _Campylobactor_ to _Campylobacter_
>   * [x]  Line 78: Add 'h' to 'Mashpit'
>   * [x]  Line 80: Correct 'calclated'
>   * [x]  Figure 1: Only parts of the figure is visible. Make sure the full figure is visible.

All the typos are corrected in the new draft.
