### 2024-12-07 #7306 @tongzhouxu
Hi @mberacochea, thank you for your feedback. We have addressed the bugs and issues raised in your comments. Below is our detailed response regarding the manuscript:

> * [x] Lines 11 and 12. PulseNet should be cited on line 11, and maybe mention that is hosted by NCBI.

We have added a PulseNet reference.

> * [x] Line 43, _Salmonella_ should be italicise
> * [x] Line 53, is sourmash an interface for Mash, or an implementation of the FracMinHash algorithm instead?

We used MinHash() in sourmash which is an interface for Mash.

> *[x] Line 80, typo "calclated"
> * [x] Figure 1, it's clipped and not readable.
> * [x] The webserver it's not mentioned in the paper.
>
> A more general note about Mashpit: If I understood correctly, the accuracy of placing a genome in the correct SNP cluster isn’t very high (70% for _Salmonella_ when considering that the "right" SNP cluster should be within the top 25). This seems quite relevant, as it’s likely important to users (this is my assumption). I would suggest expanding or rephrasing this part to make it clearer to users what the best use case for Mashpit is (it’s already mentioned in the discussion, but I feel it’s a bit lost there). I would also suggest splitting Figure 1 into two figures, one to explain the "compute performance" (in terms of resources) and the other to show the tool's accuracy.

Thank you for your suggestions! We have now split Figure 1 into two figures. Additionally, we have rephrased and expanded the performance section, specifically clarifying the best use case for Mashpit and adding an interpretation of the performance differences for _Salmonella_ and _Campylobacter_.
