### 2024-12-15 #7306 @mberacochea
Hi all,

@tongzhouxu has addressed all my comments. The software is working as expected, and the README and manuscript have been greatly improved. I second @hkaspersen in commending the work done by the authors.

I recommend that this work be published in JOSS.

Cheers
