### 2024-12-15 c5d6bf5
Minor edits to paper/bibliography
