### 2024-12-15 #7306 @csoneson
@tongzhouxu I have gone through the submission and opened a small [PR](https://github.com/tongzhouxu/mashpit/pull/32) with a few suggested edits for the manuscript and bibliography. In addition, could you please take a look at adding the missing DOIs listed above, as well as removing the unused items from the bib file? You can then generate a new proof with `@editorialbot generate pdf`
