+++ paper/paper.bib
@@ -1,15 +1,3 @@
-@article{hadfield2018nextstrain,
-  doi = {10.1093/bioinformatics/bty407},
-  title={Nextstrain: real-time tracking of pathogen evolution},
-  author={Hadfield, James and Megill, Colin and Bell, Sidney M and Huddleston, John and Potter, Barney and Callender, Charlton and Sagulenko, Pavel and Bedford, Trevor and Neher, Richard A},
-  journal={Bioinformatics},
-  volume={34},
-  number={23},
-  pages={4121--4123},
-  year={2018},
-  publisher={Oxford University Press}
-}
-
 @article{nadon2017pulsenet,
   doi = {10.2807/1560-7917.es.2017.22.23.30544},
   title={PulseNet International: Vision for the implementation of whole genome sequencing (WGS) for global food-borne disease surveillance},
@@ -95,6 +83,7 @@
 }
 
 @article{topaz2018bmscan,
+  doi={10.1186/s12879-018-3324-1},
   title={BMScan: using whole genome similarity to rapidly and accurately identify bacterial meningitis causing species},
   author={Topaz, Nadav and Boxrud, Dave and Retchless, Adam C and Nichols, Megan and Chang, How-Yi and Hu, Fang and Wang, Xin},
   journal={BMC Infectious Diseases},
@@ -105,6 +94,7 @@
 }
 
 @article{swaminathan2001pulsenet,
+  doi={10.3201/eid0703.010303},
   title={PulseNet: the molecular subtyping network for foodborne bacterial disease surveillance, {United States}},
   author={Swaminathan, Bala and Barrett, Timothy J and Hunter, Susan B and Tauxe, Robert V and {CDC PulseNet Task Force}},
   journal={Emerging Infectious Diseases},
@@ -116,6 +106,7 @@
 }
 
 @article{shu2017gisaid,
+  doi = {10.2807/1560-7917.es.2017.22.13.30494},
   title={GISAID: Global initiative on sharing all influenza data--from vision to reality},
   author={Shu, Yuelong and McCauley, John},
   journal={Eurosurveillance},
@@ -126,6 +117,7 @@
   publisher={European Centre for Disease Prevention and Control}
 }
 @article{jain2018high,
+  doi={10.1038/s41467-018-07641-9},
   title={High throughput ANI analysis of 90K prokaryotic genomes reveals clear species boundaries},
   author={Jain, Chirag and Rodriguez-R, Luis M and Phillippy, Adam M and Konstantinidis, Konstantinos T and Aluru, Srinivas},
   journal={Nature communications},
