+++ paper/paper.bib
@@ -94,7 +94,7 @@
 }
 
 @article{swaminathan2001pulsenet,
-  doi={10.3201/eid0703.010303},
+  doi={10.3201/eid0703.017303},
   title={PulseNet: the molecular subtyping network for foodborne bacterial disease surveillance, {United States}},
   author={Swaminathan, Bala and Barrett, Timothy J and Hunter, Susan B and Tauxe, Robert V and {CDC PulseNet Task Force}},
   journal={Emerging Infectious Diseases},
