### 2024-12-15 #7306 @mberacochea
> Thanks @mberacochea! For completeness, could you also tick the remaining boxes in your checklist ([#7306 (comment)](https://github.com/openjournals/joss-reviews/issues/7306#issuecomment-2395505648))?
>
> @tongzhouxu - as both reviewers are satisfied with the submission, I will also take a quick look, and will come back to you shortly with the next steps.

Done!
