### 2024-12-15 #7306 @csoneson
Thanks @mberacochea!

@tongzhouxu - the new proof looks good to me. I will make a new post below with the next steps.
