### 2024-12-15 #7306 @tongzhouxu
> ## Additional Author Tasks After Review is Complete
> * Double check authors and affiliations (including ORCIDs)
> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * Archive the release on Zenodo/figshare/etc and post the DOI here.
> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * Make sure that the license listed for the archive is the same as the software license.

Hi @csoneson , I have drafted a new release `v0.9.10` and archived the release on Zenodo at [10.5281/zenodo.14498036](https://doi.org/10.5281/zenodo.14498036). I have double checked the authors, affiliations and ORCIDs in the manuscript. I have modified the title and author list in the archive.
