### 2024-12-16 #7306 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
