### 2024-12-17 tongzhouxu/mashpit#33 @lskatz
CITATION.cff
Per [JOSS's suggestion](https://github.com/openjournals/joss-reviews/issues/7306#issuecomment-2547856695), please add CITATION.cff
