### 2024-12-17 tongzhouxu/mashpit#33 @tongzhouxu
Added CITATION.cff in ddfda49
